fix(migrations): fresh-replay — konwergencja notifications (534), TEXT id communication_plans (573), cast is_system (767), dynamiczne predykaty backupów z139 (919)
- 534: baseline tworzy minimalny notifications (257 odfiltrowana <500) —
  ADD COLUMN IF NOT EXISTS dla pełnego kształtu przed indeksami
- 573: communication_plans.id TEXT (żywy schemat ma TEXT; UUID łamał FK
  plan_id TEXT na świeżej bazie)
- 767: is_system::int = 1 (INTEGER na żywym, BOOLEAN ze świeżego 568)
- 919: backup-INSERTy budują WHERE tylko z istniejących kolumn
  (blocks_json/content_json_native istnieją tylko na żywych bazach)

+++ server/migrations/767_presentation_template_governance.sql
@@ -68,8 +68,10 @@ UPDATE presentation_templates
    SET lifecycle_state = 'approved',
        approved_at = COALESCE(approved_at, CURRENT_TIMESTAMP),
        approved_by = COALESCE(approved_by, 'system')
- -- presentation_templates.is_system is INTEGER (canonical); compare to 1 not TRUE.
- WHERE is_system = 1
+ -- is_system is INTEGER on the live schema but BOOLEAN when created fresh by
+ -- 568_presentations_brand_kits_templates.sql — cast makes the predicate work
+ -- on both (2026-07-14 fresh-replay fix; boolean::int → 1/0 in Postgres).
+ WHERE is_system::int = 1
    AND lifecycle_state = 'draft';
 
 -- Backfill lineage_root_id for every existing row so each template is
